Add copyToColumnVector in ColumnView
