+++ java/src/test/java/ai/rapids/cudf/ColumnVectorTest.java
@@ -26,6 +26,7 @@ import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -3030,9 +3031,9 @@ public class ColumnVectorTest extends CudfTestBase {
          ColumnVector expected = ColumnVector.fromLists(new HostColumnVector.ListType(true,
              new HostColumnVector.BasicType(true, DType.STRING)) , list, list3, list2)) {
       assert res1.getNullCount() == 1: "Null count is incorrect on input column";
-      assert res1.getChildColumnViewAccess(0).getNullCount() == 0 : "Null count is incorrect on input column";
+      assert res1.getChildColumnView(0).getNullCount() == 0 : "Null count is incorrect on input column";
       assert res2.getNullCount() == 0 : "Null count is incorrect on input column";
-      assert res2.getChildColumnViewAccess(0).getNullCount() == 2 : "Null count is incorrect on input column";
+      assert res2.getChildColumnView(0).getNullCount() == 2 : "Null count is incorrect on input column";
       assertColumnsAreEqual(expected, v);
     }
   }
@@ -3260,6 +3261,28 @@ public class ColumnVectorTest extends CudfTestBase {
          HostColumnVector hostColumnVector = cv.copyToHost();
          ColumnVector expected = hostColumnVector.copyToDevice()) {
       assertColumnsAreEqual(expected, cv);
+    }
+  }
+
+  @Test
+  void testCopyToColumnVector() {
+    List<Integer> list1 = Arrays.asList(10, 11, 12, 13);
+    List<Integer> list2 = Arrays.asList(16, 12, 14, 15);
+    List<Integer> list3 = Arrays.asList(0, 7, 3, 4, 2);
+
+    try(ColumnVector res = ColumnVector.fromLists(new HostColumnVector.ListType(true,
+        new HostColumnVector.BasicType(true, DType.INT32)), list1, list2, list3);
+        ColumnView childColumnView = res.getChildColumnView(0);
+        ColumnVector copiedChildCv = childColumnView.copyToColumnVector();
+        HostColumnVector hcv = copiedChildCv.copyToHost();
+        ColumnVector expected =
+            ColumnVector.fromInts(10, 11, 12, 13, 16, 12, 14, 15, 0, 7, 3, 4, 2)) {
+      System.out.println("columnView type=" + childColumnView.getType());
+      System.out.println("columnView rows=" + childColumnView.getRowCount());
+      System.out.println("copiedChildCv type=" + copiedChildCv.getType());
+      System.out.println("copiedChildCv rows=" + copiedChildCv.getRowCount());
+      System.out.println("copiedChildCv 0th=" + hcv.getInt(0));
+      System.out.println("copiedChildCv 1th=" + hcv.getInt(1));
     }
   }
 }
